ManageChiefJailer updated - error handling

+++ Final Report/proje_code_final/src/com/Group1/ManageChiefJailer.java
@@ -1,5 +1,6 @@
 package com.Group1;
 import java.util.ArrayList;
+import java.util.NoSuchElementException;
 import java.util.Queue;
 import java.util.Scanner;
 
@@ -143,6 +144,9 @@ public class ManageChiefJailer extends Jailer {
         System.out.println("Enter the ID of the jailer:");
         int jailerID = scan.nextInt();
 
+        if(dataBase.getPersonnel(jailerID) == null)
+            throw new NoSuchElementException("There is no Jailer with given id!");
+
         do {
             System.out.println("Enter the new working days:");
             shiftTimes = scan.nextLine();
@@ -218,14 +222,18 @@ public class ManageChiefJailer extends Jailer {
         int k;
         for (k = 0; k < 45; k++) System.out.print("-");
         System.out.println();
+        System.out.println("Enter the Jailer's id: ");
+        int jailerId = scan.nextInt();
+
+        if (dataBase.getPersonnel(jailerId) == null)
+            throw new NoSuchElementException("No Jailer found in given ID!");
+
         System.out.println("Enter new workplace's block: ");
         String blockName = scan.nextLine();
         Block block = new Block(blockName);
 
         System.out.println("Enter new workplace's floor: ");
         int floor = scan.nextInt();
-        System.out.println("Enter the Jailer's id: ");
-        int jailerId = scan.nextInt();
         dataBase.getPersonnel(jailerId).setPlace(new WorkPlace(block, floor));
     }
 
@@ -349,6 +357,7 @@ public class ManageChiefJailer extends Jailer {
         }
         else{
             System.out.println ("Inmate was not found!");
+            throw new NoSuchElementException("No Inmate found in given ID!");
         }
 
         dataBase.printAllVisitor();
@@ -360,11 +369,15 @@ public class ManageChiefJailer extends Jailer {
         for ( k = 0; k < 45; k++) System.out.print("-"); System.out.println();
         System.out.println("Enter Visitor's TC number: ");
         String tc = scan.nextLine();
+        if (dataBase.getVisitorWithTC(tc) == null)
+            throw new NoSuchElementException("No Visitor found in given TC!");
         removeVisitor(tc,dataBase);
     }
 
     protected void removeVisitor(String tc,DataBase dataBase)
     {
+        if (dataBase.getVisitorWithTC(tc) == null)
+            throw new NoSuchElementException("No Visitor found in given TC!");
         dataBase.deleteVisitor(new Inmate(dataBase.getVisitorWithTC(tc).inmateNumber) ,dataBase.getVisitorWithTC(tc));
     }
